requirements.old.txt
Werkzeug==1.0.1
Flask==1.1.2
Flask_Login==0.5.0
Flask_WTF==0.14.2
Flask_Script==2.0.6
Flask_Migrate==2.5.2
Flask_SQLAlchemy==2.4.0
SQLAlchemy==1.3.11
cymysql==0.9.14
WTForms==2.2.1
Authlib==0.13
requests==2.22.0
concurrent_log_handler==0.9.16
xxtea==1.3.0
behave==1.2.6
Flask-Mail==0.9.1
blinker==1.4
pycryptodome==3.9.8
